Renamed load_r.ts
